Added entry control menu component
